+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Add `project.config.defaults.enable` to turn off all default settings en masse.
   - Regressions in this PR: #169, #178
 - #175: `devShell`: Add options `benchmark` (to enable benchmark dependencies) and `hoogle` (whether to include Hoogle in development shell)
+- #189: Use `types.str` instead of `types.string` in option types
 
 ## 0.3.0 (May 22, 2023)